Added the ability to define sections.
A defined section can be created within a tempate and then reused within
the template.  The content is directly rendered to the defined section
and when the section is used, the already-rendered content is
substituted in place.  This is not the same as a macro.  Basically, it
defines a named buffer that all rendered content gets redirected to, and
can be used to render from.  It also allows an application to make
special use to those defined sections in whatever way they see fit.
Currently, since the parse doesn't handle strings it is of limited use
until the parser is revamped.