Included files can now use expessions.
I've added expression support to included files.  Also included files
are not loaded at parse time, but run-time because of this.  It can
present a security feature if templates come from an untrusted source,
or if untrused user input is part of the include expression.  This can
be mitigated by using a loader that restricts from where templates can
be loaded.