Changed how scopes work.
The environment has a scope stack.  Entering a template pushes
the stack.  Variables are accessed from the closest scope up.
This allows a template to be recursive if needed.  However,
there is currently nothing that limits this recursion and so
care must be take with the templates that are recurisve.