Added extra characters for line control.
If a tag starts with "-" then it will eat up to and including and entire
new line still.  However, if it starts with "<", "!", or "^", it will
each up to but not including the newline.  The extra characters were
added to allow for syntax highlighters that would treat "<" as a tag.