Added callback support.
A script can register a callback function that can be called from the
template as follows:

    {% callback name %}
    {% callback name(params, more_params) %}

This function will receive the environment, the source filename of the
script, and a list of evaluated parameters.  If the callback returns
anything other than None, the string value of that will be rendered to
the renderer.