Added import feature to create importable libraries.
The import feature simply allows a library to be registered with a
unique name, then for each template, it can be imported to any specific
variable.  This allows to a template to decide what name the library
should use in itself instead of needing to set it as a global to a
specific variable on environment creation and the templates being forced
to use that variable throughout.