Added ability to search the root paths for a template.
If a root path or list of root paths is specified, then a template can
be loaded by searching the root paths in order by specifyng a filename
in the following format:

    @path/to/file.tmpl

In any other form, normal loading behaviour will be used, based on the
current directory or parent template path.  This feature can not be used
if no roots are specified.  Any template loaded must be under one of the
roots.