Changed handling of includes.
Each tempalte has the filename and relname.  If roots are not used in
the template loader, the relname isn't used.

When including a template, if can be prefixed with "@/" or "@".  If it
is prefixed with "@/", then it specfies a relative name to search from
the path roots.  If "@" without a leading "/", then a relative name
tosearch, relative to the parents relative name, over the root paths.
Once a file is found, both the relname and filename is stored internally
in the template. These can only be used if roots are used.

If it is not prefixed with "@", then it specifies a file path relative
to the parent filename.  It is then checked against the roots if root
paths are used, and a relative name is determined.  If roots are not
used, then a relative name is not determines.

So, basicaly, if it begins with "@/", search for the template based on
the name over the root paths.  If is begins with "@", take the relative
name of the current template, join it with the specified name, and
search for that over the root paths.  Otherwise, search relative to the
filename and make sure it is in the root paths.