Removed push_autostrip and replaced it with strip/endstrip
The reason was due to the fact that the code tag also pushed and popped
the value of autostrip.  This would cause a problem if autostrip was
pushed or popped inside the code block unless properly nested. Using a
strip block and adding it to the ops stack ensures proper nesting
However the strip value should not be changed in a code block.