Changed variable access approach.
Variables access as "name.subname" will first attempt to look up subname
as an item of the variable name.  If it does not exist, it will then
attempt to lookup up subname as an attribute of the variable name. To
force item lookup, use "name.#subname".  To force attribute lookup use
"name.@subname".

The variable lookup no longer uses prefixed "attr_" or "func_" to
determine how to handle a lookup. To result in a variable lookup calling
a function without the template making the call, ensure that the method
is made a propety using the "@property" decorator. Then when the value
is retrieve in "Environment.get" with getattr, it will automatically
call the function.