Modified service container mixin.
For the service container, a service is explicitly specified as a
factory or a singleton based on whether it is registered as
register_factory or register_singleton.  When calling resolve_service,
arguments can be passed if the servivce is a factory allowing each call
to create a new instance.  Arguments should not be passed if it is a
singleton, as only the first call creates the instance while futher
calls simple return the existing instance.